# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales Support API",
        "description": "SES企業向け営業支援システム API",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "ローカル開発環境"
        }
    ],
    "paths": {
...
        "/api/v1/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "担当者一覧取得",
                "operationId": "b96f6b9663eca5d37c581f0cf15c471d",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "氏名・部署・役職・会社名で検索",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "顧客IDで絞り込み",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "ページ番号",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "担当者登録",
                "operationId": "bf0b491f5df1cc8cdb0995ea168b9ccd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "name"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxx-xxxx"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "山田 太郎"
                                    },
                                    "department": {
                                        "type": "string",
                                        "example": "営業部"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "部長"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "yamada@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "03-1234-5678"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "備考"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "登録成功"
                    },
                    "422": {
                        "description": "バリデーションエラー"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/contacts/{id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "担当者詳細取得",
                "operationId": "627b088f95647a22776924ddebefe7de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "担当者ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "404": {
                        "description": "担当者が見つかりません"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "担当者更新",
                "operationId": "7f57e34a4cc2a52acf2c4355ee2b42d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "担当者ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "name"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "山田 太郎"
                                    },
                                    "department": {
                                        "type": "string",
                                        "example": "営業部"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "部長"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "yamada@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "03-1234-5678"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    },
                    "422": {
                        "description": "バリデーションエラー"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "担当者削除",
                "operationId": "36c5deaf97d3489715bbeb0d3f13fc03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "担当者ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "削除成功"
                    },
                    "404": {
                        "description": "担当者が見つかりません"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Supabase Auth JWT トークン",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}